PullRequest: graalpython/3229

+++ graalpython/com.oracle.graal.python/src/com/oracle/graal/python/builtins/objects/cext/hpy/jni/GraalHPyJNIContext.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019, 2023, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2019, 2024, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * The Universal Permissive License (UPL), Version 1.0
@@ -1677,7 +1677,7 @@ public final class GraalHPyJNIContext extends GraalHPyNativeContext {
         } else {
             Object object = context.getObjectForHPyHandle(GraalHPyBoxing.unboxHandle(handle));
             try {
-                return (double) PyLongAsDoubleNode.executeUncached(object);
+                return PyLongAsDoubleNode.executeUncached(object);
             } catch (PException e) {
                 HPyTransformExceptionToNativeNode.executeUncached(context, e);
                 return -1L;
